+++ Dyna/docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Reversed the Burst Charge card upgrades.
+
 ## 1.0.0
 Released 15 April 2024.